# Performance improvement for elasticities