It's a bit more efficient to split a data.table and then call a specific element in the loop rather than filtering the whole data.table at each iteration.

Most of the improvement comes from the change to `merge()`, but I really can't tell why it's more efficient, it just seems to be.

### Benchmarks
(I changed a bit my way of benchmarking, now all results are in the table at the end)

```r
library(bench)

# run same benchmark both on 'main' and on my fork
# https://github.com/DavisVaughan/cross
out <- cross::run(
  pkgs = c("vincentarelbundock/marginaleffects", 
           "etiennebacher/marginaleffects@speedup-elasticities"),
  ~ {
    library(marginaleffects)
    
    # 30,000 obs, 23 predictors
    N <- 30000
    dat <- data.frame(matrix(rnorm(N * 24), ncol = 24))
    mod <- lm(X1 ~ ., dat)
    
    bench::mark(
      avg_slopes(mod, slope = "eydx"), 
      avg_slopes(mod, slope = "dydx"), 
      avg_slopes(mod, slope = "eyex"), 
      check = FALSE,
      iterations = 10
    )
  }
)

tidyr::unnest(out, result) |> 
  dplyr::select(pkg, expression, min, median, mem_alloc) |> 
  dplyr::mutate(pkg = ifelse(grepl("vinc", pkg), "main", "fork")) |> 
  dplyr::arrange(expression)
#> # A tibble: 6 × 5
#>   pkg   expression                               min   median mem_alloc
#>   <chr> <bch:expr>                          <bch:tm> <bch:tm> <bch:byt>
#> 1 main  "avg_slopes(mod, slope = \"eydx\")"   34.75s   41.69s    29.5GB
#> 2 fork  "avg_slopes(mod, slope = \"eydx\")"   25.37s    26.9s   21.84GB
#> 3 main  "avg_slopes(mod, slope = \"dydx\")"    4.79s    4.93s    6.05GB
#> 4 fork  "avg_slopes(mod, slope = \"dydx\")"    4.32s     4.6s    6.05GB
#> 5 main  "avg_slopes(mod, slope = \"eyex\")"   32.83s   34.62s   29.62GB
#> 6 fork  "avg_slopes(mod, slope = \"eyex\")"   24.08s   24.81s   21.96GB
```