@etiennebacher 2023-05-23
@vincentarelbundock wait before reviewing/merging, I can add another speedup later