@vincentarelbundock 2023-05-23
I'm a bit worried about this PR. Modifying in-place guarantees that we preserve row order. I'm not sure if splitting may reorder rows sometimes depending on the splitting conditions and the many different types of objects produced by the package.

I don't have time to think about this seriously now, but I've had several row order bugs recently, so that's on my mind right now.