@etiennebacher 2023-05-23
I removed the use of `split()` to come back to the original filtering, this was only a minor change in speed and memory anyway. I made an additional speed improvement by moving some code outside the loop. 

I don't think I'll make other changes to this, but let me know if you want to add more tests for row order.

``` r
library(bench)

out <- cross::run(
  pkgs = c("vincentarelbundock/marginaleffects",
           "etiennebacher/marginaleffects@speedup-elasticities"),
  ~ {
    library(marginaleffects)
    
    N <- 30000
    dat <- data.frame(matrix(rnorm(N * 24), ncol = 24))
    mod <- lm(X1 ~ ., dat)
    
    bench::mark(
      avg_slopes(mod, slope = "eydx"), 
      avg_slopes(mod, slope = "dydx"), 
      avg_slopes(mod, slope = "eyex"), 
      check = FALSE,
      iterations = 10
    )
  }
)

tidyr::unnest(out, result) |> 
  dplyr::select(pkg, expression, min, median, mem_alloc) |> 
  dplyr::mutate(
    pkg = ifelse(grepl("vinc", pkg), "main", "fork")
  ) |> 
  dplyr::arrange(expression)
#> # A tibble: 6 × 5
#>   pkg   expression                               min   median mem_alloc
#>   <chr> <bch:expr>                          <bch:tm> <bch:tm> <bch:byt>
#> 1 main  "avg_slopes(mod, slope = \"eydx\")"   31.87s   33.26s    29.5GB
#> 2 fork  "avg_slopes(mod, slope = \"eydx\")"   19.61s    20.1s   14.11GB
#> 3 main  "avg_slopes(mod, slope = \"dydx\")"    4.15s    4.72s    6.05GB
#> 4 fork  "avg_slopes(mod, slope = \"dydx\")"     5.1s    5.46s    6.05GB
#> 5 main  "avg_slopes(mod, slope = \"eyex\")"   32.52s   34.84s   29.62GB
#> 6 fork  "avg_slopes(mod, slope = \"eyex\")"   21.43s   24.27s   14.23GB
```