@vincentarelbundock 2023-05-23
That's great, thanks!

I'd like to run the example on my machine to see if the 20% loss on `dydx` was an artifact or a real loss. Where's is the `cross` package from?